profile page, fixing issues and changes. Made by Filipe Santos (53304)
added functioning profile page with the possibility to change profile info and the option to delete the current account.
correction of error in complete register where the the form wouldnt appear if user registered as a supplier. changes to complete register in accordance to changes made in user_address table in th database.
Made by Filipe Santos (53304)

Co-Authored-By: fc53304 <80914551+foolemp@users.noreply.github.com>